documents/documentsProcessor.py
from PIL import Image
import pytesseract


def extract_text_from_document(file_path):
    try:
        image = Image.open(file_path)
        text = pytesseract.image_to_string(image)
        return text
    except Exception:
        return ""
