requirements.txt
flask
pillow
pytesseract
python-dotenv
google-genai
pymupdf
sqlite-utils
